refactor: some refactor
