+++ import-service/src/functions/BlobImportProductsFromFile.test.ts
@@ -30,4 +30,19 @@ Age of Wonders 4,"Rule a fantasy realm!",35.99,100`,
       })
     );
   });
+
+  test("parse empty csv file", async () => {
+    const context = new InvocationContext({
+      functionName: "HttpGetImportProductsFiles",
+      triggerMetadata: {
+        name: "products.csv",
+      },
+    });
+    const logSpy = jest.spyOn(context, "log");
+
+    const buffer = Buffer.from("", "utf8");
+    await BlobImportProductsFromFile(buffer, context);
+
+    expect(logSpy).toHaveBeenNthCalledWith(2, "No items found.");
+  });
 });
